+++ internal/node/keyring_linux_test.go
@@ -1,0 +1,15 @@
+//go:build linux
+
+package node
+
+import (
+	"os"
+	"testing"
+
+	"entropy/internal/vault"
+)
+
+func TestMain(m *testing.M) {
+	vault.UseMemoryKeyringForTests()
+	os.Exit(m.Run())
+}